add: textures; fix: debug menu draws in 3D space, wip: chunks
Added some new textures. Fixed the issue with the giant floating FPS counter in 3D space. Got some work done with chunks (see chunk.v:3)

+++ src/util.v
@@ -3,7 +3,7 @@ import sokol.sgl
 import transform { Vector3 }
 
 // sgl_draw_cube draws a cube to the scren using the sokol.sgl module.
-fn sgl_draw_cube(size int) {
+fn sgl_draw_cube(size f32) {
 	// vfmt off
 	sgl.begin_quads()
 	{
